backend/requirements.txt
flask
flask-sqlalchemy
werkzeug
pyjwt